Backend/requirements.txt
django===3.2
djangorestframework===3.14.0
psycopg2-binary